// back_end/src/auth/type.ts
// export interface StoreAuthToken {
//     accessToken: string;
//     refreshToken: string;
//     grayAccessToken: string;
//     blackList: string[];
// }

export interface Store_Auth_Token_Field {
    access_token: string;
    refresh_token: string;
    gray_access_token: string;
    black_list: string[];
}